tests: Move evaluation_config fixture to global conftest.py
